msg-alert-delete-user

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\User;
+use App\Call;
+use App\Contact;
 
 class UserController extends Controller
 {
@@ -13,7 +15,15 @@ class UserController extends Controller
     }
     public function destroyUser(Request $request){
     	$user = User::find($request->id);
-    	$user->delete();
-    	return redirect()->back()->with('alert','Registro Apagado');
+
+        $call    = Call::where('user_id',$request->id)->count();
+        $contact = Contact::where('user_id',$request->id)->count();
+
+        if($call == 0 && $contact == 0){
+	    	$user->delete();
+	    	return redirect()->back()->with('alert','Registro Apagado');
+	    }else{
+	    	return redirect()->back()->with('alert','Não foi possível apagar existe ligações ou contatos vinculados');
+	    }
     }
 }
